Renamed CacheManager to CachedSettingsManager
