+++ .travis.yml
@@ -9,10 +9,10 @@ php:
   - hhvm
 
 before_install:
-  - composer self-update
+  - travis_retry composer self-update
 
 install:
-  - composer install
+  - composer install --ignore-platform-reqs
 
 script:
   - phpunit
