+++ composer.json
@@ -26,7 +26,7 @@
         "symfony/form": "~2.1,!=2.5.0",
         "symfony/validator": "~2.1,!=2.5.0",
         "symfony/translation": "~2.1,!=2.5.0",
-        "doctrine/orm": "~2.2,>=2.2.3,<2.5"
+        "doctrine/orm": "~2.2,>=2.2.3"
     },
     "require-dev": {
         "phpunit/phpunit": "~4.6",
